Fix Drive Scope; Add Gmail
Fix the list of Authorization Scopes to include all of Drive and Gmail
Compose. This resolves an issue with getting the meta-data of a Drive
file and adds sending email through Gmail.

This removes the need to have a mail.dat file or any extra credential
files. Only the Client ID JSON from Google is required now.

No changes were made to the format of the email itself.